pyproject.toml
[tool.poetry]
name = "paper-animator"
version = "0.0.1"
description = "Animate a version-controlled PDF/LaTeX paper!"
authors = ["sneakers-the-rat <JLSaunders987@gmail.com>"]
license = "MPL-2.0"
homepage = "https://github.com/sneakers-the-rat/paper-animator"
repository = "https://github.com/sneakers-the-rat/paper-animator"

[tool.poetry.dependencies]
python = "^3.8"
gitpython = "^3.1.13"
pymupdf = "^1.18.8"
tqdm = "^4.57.0"
numpy = "^1.20.1"
matplotlib = "^3.3.4"
scikit-video = "^1.1.11"
pypng = "^0.0.20"

[tool.poetry.dev-dependencies]

# New scripts
[tool.poetry.scripts]
paper-animator = 'paper_animator:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
